requirements.txt
cachetools==5.5.2
datasets==2.20.0
einops==0.8.1
matplotlib==3.9.4
numpy==1.24.4
pandas==2.2.3
Pillow==11.1.0
pycocoevalcap==1.2
python_dateutil==2.9.0.post0
pytorch_lightning==2.3.0
PyYAML==6.0.1
radgraph==0.0.9
scikit_learn==1.6.1
scipy==1.11.2
sentence_transformers==3.0.1
# torch==2.3.1+cu118
torchmetrics==1.4.0.post0
# torchvision==0.18.1+cu118
tqdm==4.66.4
transformers==4.43.3